Fix multisite bugs & enhance functionality
Bugs fixed:
- saving of settings on the multisite settings page was not working.
- restoring site to default settings from multisite settings page was not working.
- initializing new blogs with settings from a chosen default blog was not working.
- wrong option debug information shown on multisite settings page

Enhancements:
- When using the 'Reset default settings' button on a blog in a network while another blog has been chosen to be used as a basis for the settings for all new blogs, the reset will respect that setting and reset the blog to the settings from the chosen blog.
- For small networks ( < 100 sites ), the network page user interface has been improved, by offering drop-down lists of the blogs for blog selection fields. For larger networks, the interface remains the same.

Also:
- Added a few extra options to the network blog reset/copy exclude list (theme dependent ones)
- clean up wrongly stored options (i.e. wpseo_ms in options table not in sitemeta table)
- made the padding on select boxes slightly smaller as letters were being cut off at the bottom.

+++ inc/class-wpseo-options.php
@@ -33,8 +33,9 @@ if ( ! class_exists( 'WPSEO_Option' ) ) {
 	 *    Oh and the very few options where the default value is null, i.e. wpseo->'theme_has_description'
 	 *
 	 * [Updating/Adding options]
-	 * - Use the normal add/update_option() functions. As long a the classes here are instantiated, validation
-	 *    for all options and their subkeys will be automatic.
+	 * - For multisite site_options, please use the WPSEO_Options::update_site_option() method.
+	 * - For normal options, use the normal add/update_option() functions. As long a the classes here
+	 *   are instantiated, validation for all options and their subkeys will be automatic.
 	 * - On (succesfull) update of a couple of options, certain related actions will be run automatically.
 	 *    Some examples:
 	 *      - on change of wpseo[yoast_tracking], the cron schedule will be adjusted accordingly
@@ -134,15 +135,29 @@ if ( ! class_exists( 'WPSEO_Option' ) ) {
 			$this->add_default_filters(); // return defaults if option not set
 			$this->add_option_filters(); // merge with defaults if option *is* set
 
-			/* The option validation routines remove the default filters to prevent failing
-			   to insert an option if it's new. Let's add them back afterwards. */
-			add_action( 'add_option', array( $this, 'add_default_filters' ) ); // adding back after INSERT
 
-			if ( version_compare( $GLOBALS['wp_version'], '3.7', '!=' ) ) { // adding back after non-WP 3.7 UPDATE
-				add_action( 'update_option', array( $this, 'add_default_filters' ) );
-			} else { // adding back after WP 3.7 UPDATE
-				add_filter( 'pre_update_option_' . $this->option_name, array( $this, 'wp37_add_default_filters' ) );
+			if ( $this->multisite_only !== true ) {
+				/* The option validation routines remove the default filters to prevent failing
+				   to insert an option if it's new. Let's add them back afterwards. */
+				add_action( 'add_option', array( $this, 'add_default_filters' ) ); // adding back after INSERT
+
+				if ( version_compare( $GLOBALS['wp_version'], '3.7', '!=' ) ) { // adding back after non-WP 3.7 UPDATE
+					add_action( 'update_option', array( $this, 'add_default_filters' ) );
+				} else { // adding back after WP 3.7 UPDATE
+					add_filter( 'pre_update_option_' . $this->option_name, array( $this, 'wp37_add_default_filters' ) );
+				}
 			}
+			else if ( is_multisite() ) {
+				/* The option validation routines remove the default filters to prevent failing
+				   to insert an option if it's new. Let's add them back afterwards.
+				   
+				   For site_options, this method is not foolproof as these actions are not fired
+				   on an insert/update failure. Please use the WPSEO_Options::update_site_option() method
+				   for updating site options to make sure the filters are in place. */
+				add_action( 'add_site_option_' . $this->option_name, array( $this, 'add_default_filters' ) );
+				add_action( 'update_site_option_' . $this->option_name, array( $this, 'add_default_filters' ) );
+			}
+
 
 
 			/* Make sure the option will always get validated, independently of register_setting()
@@ -208,9 +223,6 @@ if ( ! class_exists( 'WPSEO_Option' ) ) {
 			if ( has_filter( 'default_option_' . $this->option_name, array( $this, 'get_defaults' ) ) === false ) {
 				add_filter( 'default_option_' . $this->option_name, array( $this, 'get_defaults' ) );
 			}
-			if ( has_filter( 'default_site_option_' . $this->option_name, array( $this, 'get_defaults' ) ) === false ) {
-				add_filter( 'default_site_option_' . $this->option_name, array( $this, 'get_defaults' ) );
-			}
 		}
 
 
@@ -238,7 +250,6 @@ if ( ! class_exists( 'WPSEO_Option' ) ) {
 		 */
 		public function remove_default_filters() {
 			remove_filter( 'default_option_' . $this->option_name, array( $this, 'get_defaults' ) );
-			remove_filter( 'default_site_option_' . $this->option_name, array( $this, 'get_defaults' ) );
 		}
 
 
@@ -271,9 +282,6 @@ if ( ! class_exists( 'WPSEO_Option' ) ) {
 			if ( has_filter( 'option_' . $this->option_name, array( $this, 'get_option' ) ) === false ) {
 				add_filter( 'option_' . $this->option_name, array( $this, 'get_option' ) );
 			}
-			if ( has_filter( 'site_option_' . $this->option_name, array( $this, 'get_option' ) ) === false ) {
-				add_filter( 'site_option_' . $this->option_name, array( $this, 'get_option' ) );
-			}
 		}
 
 
@@ -285,7 +293,6 @@ if ( ! class_exists( 'WPSEO_Option' ) ) {
 		 */
 		public function remove_option_filters() {
 			remove_filter( 'option_' . $this->option_name, array( $this, 'get_option' ) );
-			remove_filter( 'site_option_' . $this->option_name, array( $this, 'get_option' ) );
 		}
 
 
@@ -346,8 +353,13 @@ if ( ! class_exists( 'WPSEO_Option' ) ) {
 
 
 			$option_value = array_map( array( __CLASS__, 'trim_recursive' ), $option_value );
-			$old          = get_option( $this->option_name );
-			$clean        = $this->validate_option( $option_value, $clean, $old );
+			if ( $this->multisite_only !== true ) {
+				$old = get_option( $this->option_name );
+			}
+			else {
+				$old = get_site_option( $this->option_name );
+			}
+			$clean = $this->validate_option( $option_value, $clean, $old );
 
 			/* Retain the values for variable array keys even when the post type/taxonomy is not yet registered */
 			if ( isset( $this->variable_array_key_patterns ) ) {
@@ -367,7 +379,7 @@ if ( ! class_exists( 'WPSEO_Option' ) ) {
 		abstract protected function validate_option( $dirty, $clean, $old );
 
 
-		/* *********** METHODS for ADDING/UPGRADING the option *********** */
+		/* *********** METHODS for ADDING/UPDATING/UPGRADING the option *********** */
 		
 		/**
 		 * Retrieve the real old value (unmerged with defaults)
@@ -377,7 +389,15 @@ if ( ! class_exists( 'WPSEO_Option' ) ) {
 		protected function get_original_option() {
 			$this->remove_default_filters();
 			$this->remove_option_filters();
-			$option_value = get_option( $this->option_name ); // = (unvalidated) array, NOT merged with defaults
+
+			// Get (unvalidated) array, NOT merged with defaults
+			if ( $this->multisite_only !== true ) {
+				$option_value = get_option( $this->option_name );
+			}
+			else {
+				$option_value = get_site_option( $this->option_name );
+			}
+
 			$this->add_option_filters();
 			$this->add_default_filters();
 			
@@ -393,7 +413,37 @@ if ( ! class_exists( 'WPSEO_Option' ) ) {
 		 */
 		public function maybe_add_option() {
 			if ( $this->get_original_option() === false ) {
-				update_option( $this->option_name, $this->get_defaults() );
+				if ( $this->multisite_only !== true ) {
+					update_option( $this->option_name, $this->get_defaults() );
+				}
+				else {
+					$this->update_site_option( $this->get_defaults() );
+				}
+			}
+		}
+
+
+		/**
+		 * Update a site_option
+		 *
+		 * @internal This special method is only needed for multisite options, but very needed indeed there.
+		 * The order in which certain functions and hooks are run is different between get_option() and
+		 * get_site_option() which means in practice that the removing of the default filters would be
+		 * done too late and the re-adding of the default filters might not be done at all.
+		 * Aka: use the WPSEO_Options::update_site_option() method (which calls this method) for
+		 * safely adding/updating multisite options.
+		 *
+		 * @return bool whether the update was succesfull
+		 */
+		public function update_site_option( $value ) {
+			if ( $this->multisite_only === true && is_multisite() ) {
+				$this->remove_default_filters();
+				$result = update_site_option( $this->option_name, $value );
+				$this->add_default_filters();
+				return $result;
+			}
+			else {
+				return false;
 			}
 		}
 
@@ -445,7 +495,12 @@ if ( ! class_exists( 'WPSEO_Option' ) ) {
 
 			/* Save the cleaned value - validation will take care of cleaning out array keys which
 			   should no longer be there */
-			update_option( $this->option_name, $option_value );
+			if ( $this->multisite_only !== true ) {
+				update_option( $this->option_name, $option_value );
+			}
+			else {
+				$this->update_site_option( $this->option_name, $option_value );
+			}
 		}
 
 
@@ -2782,8 +2837,17 @@ if ( ! class_exists( 'WPSEO_Option_Social' ) ) {
 /*******************************************************************
  * Option: wpseo_ms
  *******************************************************************/
-if ( ! class_exists( 'WPSEO_Option_MS' ) ) {
+if ( is_multisite() && ! class_exists( 'WPSEO_Option_MS' ) ) {
 
+	/**
+	 * Site option for Multisite installs only
+	 *
+	 * This class will not even be available/loaded if not on multisite, so none of the actions will
+	 * register if not on multisite.
+	 *
+	 * Overloads a number of methods of the abstract class to ensure the use of the correct site_option
+	 * WP functions.
+	 */
 	class WPSEO_Option_MS extends WPSEO_Option {
 
 		/**
@@ -2844,20 +2908,54 @@ if ( ! class_exists( 'WPSEO_Option_MS' ) ) {
 
 
 		/**
-		 * Register (whitelist) the option for the configuration pages.
-		 * The validation callback is already registered separately on the sanitize_option hook,
-		 * so no need to double register.
+		 * Add filters to make sure that the option default is returned if the option is not set
 		 *
-		 * @todo [JRF] if the extra bit below is no longer needed, move the if combined with a check
-		 * for $this->multisite_only to the abstract class
-		 *
-		 * @return void
+		 * @return  void
 		 */
-		public function register_setting() {
-			if ( is_multisite() && WPSEO_Options::grant_access() ) {
-				register_setting( $this->group_name, $this->option_name );
+		public function add_default_filters() {
+			// Don't change, needs to check for false as could return prio 0 which would evaluate to false
+			if ( has_filter( 'default_site_option_' . $this->option_name, array( $this, 'get_defaults' ) ) === false ) {
+				add_filter( 'default_site_option_' . $this->option_name, array( $this, 'get_defaults' ) );
 			}
 		}
+
+
+		/**
+		 * Remove the default filters.
+		 * Called from the validate() method to prevent failure to add new options
+		 *
+		 * @return  void
+		 */
+		public function remove_default_filters() {
+			remove_filter( 'default_site_option_' . $this->option_name, array( $this, 'get_defaults' ) );
+		}
+
+
+		/**
+		 * Add filters to make sure that the option is merged with its defaults before being returned
+		 *
+		 * @return  void
+		 */
+		public function add_option_filters() {
+			// Don't change, needs to check for false as could return prio 0 which would evaluate to false
+			if ( has_filter( 'site_option_' . $this->option_name, array( $this, 'get_option' ) ) === false ) {
+				add_filter( 'site_option_' . $this->option_name, array( $this, 'get_option' ) );
+			}
+		}
+
+
+		/**
+		 * Remove the option filters.
+		 * Called from the clean_up methods to make sure we retrieve the original old option
+		 *
+		 * @return  void
+		 */
+		public function remove_option_filters() {
+			remove_filter( 'site_option_' . $this->option_name, array( $this, 'get_option' ) );
+		}
+
+
+		/* *********** METHODS influencing add_uption(), update_option() and saving from admin pages *********** */
 
 
 		/**
@@ -2888,7 +2986,7 @@ if ( ! class_exists( 'WPSEO_Option_MS' ) ) {
 
 
 					case 'defaultblog':
-						if ( isset( $dirty[ $key ] ) && $dirty[ $key ] !== '' ) {
+						if ( isset( $dirty[ $key ] ) && ( $dirty[ $key ] !== '' && $dirty[ $key ] !== '-' ) ) {
 							$int = self::validate_int( $dirty[ $key ] );
 							if ( $int !== false && $int > 0 ) {
 								// Check if a valid blog number has been received
@@ -3417,7 +3515,12 @@ if ( ! class_exists( 'WPSEO_Options' ) ) {
 		 */
 		protected function __construct() {
 			foreach ( self::$options as $option_name => $option_class ) {
-				self::$option_instances[ $option_name ] = call_user_func( array( $option_class, 'get_instance' ) );
+				if ( class_exists( $option_class ) ) {
+					self::$option_instances[ $option_name ] = call_user_func( array( $option_class, 'get_instance' ) );
+				}
+				else {
+					unset( self::$options[ $option_name ] );
+				}
 			}
 		}
 
@@ -3496,6 +3599,25 @@ if ( ! class_exists( 'WPSEO_Options' ) ) {
 
 			return null;
 		}
+		
+		
+		/**
+		 * Update a site_option
+		 *
+		 * @param  string $option_name The option name of the option to save
+		 * @param  mized  $value       The new value for the option
+		 *
+		 * @return bool
+		 */
+		public static function update_site_option( $option_name, $value ) {
+			if ( is_network_admin() && isset( self::$option_instances[ $option_name ] ) ) {
+				return self::$option_instances[ $option_name ]->update_site_option( $value );
+			}
+			else {
+				return false;
+			}
+		}
+
 
 		/**
 		 * Get the instantiated option instance
@@ -3550,7 +3672,13 @@ if ( ! class_exists( 'WPSEO_Options' ) ) {
 
 			if ( is_array( $option_names ) && $option_names !== array() ) {
 				foreach ( $option_names as $option_name ) {
-					$option = get_option( $option_name );
+					if ( self::$option_instances[ $option_name ]->multisite_only !== true ) {
+						$option = get_option( $option_name );
+					}
+					else {
+						$option = get_site_option( $option_name );
+					}
+
 					if ( is_array( $option ) && $option !== array() ) {
 						$all_options = array_merge( $all_options, $option );
 					}
@@ -3644,19 +3772,24 @@ if ( ! class_exists( 'WPSEO_Options' ) ) {
 		 *
 		 * @static
 		 * @return void
-		 *
-		 * @todo [JRF => Yoast] add check for multisite and only add multisite option if applicable ?
-		 * - currently will not add it (=old behaviour)
-		 * @todo [JRF => Yoast] may be check for default blog option if multisite and restore based on
-		 * that if available ?
 		 */
 		public static function reset() {
-			$option_names = self::get_option_names();
-			if ( is_array( $option_names ) && $option_names !== array() ) {
-				foreach ( $option_names as $option_name ) {
-					delete_option( $option_name );
-					update_option( $option_name, get_option( $option_name ) );
+			if ( ! is_multisite() ) {
+				$option_names = self::get_option_names();
+				if ( is_array( $option_names ) && $option_names !== array() ) {
+					foreach ( $option_names as $option_name ) {
+						delete_option( $option_name );
+						update_option( $option_name, get_option( $option_name ) );
+					}
 				}
+			}
+			else {
+				// Reset MS blog based on network default blog setting
+				self::reset_ms_blog( get_current_blog_id() );
+
+				$option                    = get_option( 'wpseo' ); // Renew option after reset
+				$option['ms_defaults_set'] = true;
+				update_option( 'wpseo', $option );
 			}
 
 			self::initialize();
@@ -3676,8 +3809,7 @@ if ( ! class_exists( 'WPSEO_Options' ) ) {
 			$option = get_option( 'wpseo' );
 
 			if ( is_multisite() && $option['ms_defaults_set'] === false ) {
-				$current_site = get_current_site();
-				self::reset_ms_blog( $current_site->id );
+				self::reset_ms_blog( get_current_blog_id() );
 
 				$option                    = get_option( 'wpseo' ); // Renew option after reset
 				$option['ms_defaults_set'] = true;
@@ -3701,43 +3833,52 @@ if ( ! class_exists( 'WPSEO_Options' ) ) {
 		 * @return  void
 		 */
 		public static function reset_ms_blog( $blog_id ) {
-			$options      = get_site_option( 'wpseo_ms' );
-			$option_names = self::get_option_names();
-			// All of these settings shouldn't be copied so easily
-			// @todo [JRF => Yoast] Would it be an idea to move and split this to be a class property of each option class, so as to keep the information contained in the most logical place ?
-			// Let me know and I'll implement.
-			$exclude = array(
-				'ignore_blog_public_warning',
-				'ignore_meta_description_warning',
-				'ignore_page_comments',
-				'ignore_permalink',
-				'ignore_tour',
-				'theme_description_found',
-				'theme_has_description',
-				'alexaverify',
-				'googleverify',
-				'msverify',
-				'pinterestverify',
-				'yandexverify',
-				'fb_admins',
-				'fbapps',
-				'fbconnectkey',
-				'fbadminapp',
-			);
+			if ( is_multisite() ) {
+				$options      = get_site_option( 'wpseo_ms' );
+				$option_names = self::get_option_names();
+				// All of these settings shouldn't be copied so easily
+				// @todo [JRF => Yoast] Would it be an idea to move and split this to be a class property of each option class, so as to keep the information contained in the most logical place ?
+				// Let me know and I'll implement.
+				$exclude = array(
+					'ignore_blog_public_warning',
+					'ignore_meta_description_warning',
+					'ignore_page_comments',
+					'ignore_permalink',
+					'ignore_tour',
+					
+					/* theme dependent */
+					'title_test',
+					'forcerewritetitle',
+					'theme_description_found',
+					'theme_has_description',
 
-			if ( is_array( $option_names ) && $option_names !== array() ) {
-				$base_blog_id = $blog_id;
-				if ( $options['defaultblog'] !== '' && $options['defaultblog'] != 0 ) {
-					$base_blog_id = $options['defaultblog'];
-				}
-				foreach ( $option_names as $option_name ) {
-					delete_blog_option( $blog_id, $option_name );
-
-					$new_option = get_blog_option( $base_blog_id, $option_name );
-					foreach ( $exclude as $key ) {
-						unset( $new_option[ $key ] );
+					/* privacy */
+					'alexaverify',
+					'googleverify',
+					'msverify',
+					'pinterestverify',
+					'yandexverify',
+					'fb_admins',
+					'fbapps',
+					'fbconnectkey',
+					'fbadminapp',
+				);
+	
+				if ( is_array( $option_names ) && $option_names !== array() ) {
+					$base_blog_id = $blog_id;
+					if ( $options['defaultblog'] !== '' && $options['defaultblog'] != 0 ) {
+						$base_blog_id = $options['defaultblog'];
 					}
-					update_blog_option( $blog_id, $option_name, $new_option );
+
+					foreach ( $option_names as $option_name ) {
+						delete_blog_option( $blog_id, $option_name );
+	
+						$new_option = get_blog_option( $base_blog_id, $option_name );
+						foreach ( $exclude as $key ) {
+							unset( $new_option[ $key ] );
+						}
+						update_blog_option( $blog_id, $option_name, $new_option );
+					}
 				}
 			}
 		}
